Add files via upload
The two scripts here are for a workflow that leverages a segmenting CNN with recurrent units for applications of  segmenting three-dimension volumes (think MRI scans). The workflow involves stage-wise training where the U-Net architecture is first trained on randomly sampled data. Then, this trained architecture is augmented with two convolutional LSTM layers that are trained on adjacent slices of the training data in order to learn to leverage intra-slice context.